refactor: フォルダ構成とエントリポイントを変更
祝日判定関数の実装（#8）に向けた準備として、以下を変更した:

- ファイル命名規則をケバブケースからキャメルケースに変更
- フォルダ構成を _internal/ と _data/ に移行
- エントリポイントを変更（light.ts → index.ts, index.ts → full.ts）
- package.json の exports を更新
