+++ package.json
@@ -16,9 +16,9 @@
       "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
     },
-    "./light": {
-      "types": "./dist/light.d.ts",
-      "default": "./dist/light.js"
+    "./full": {
+      "types": "./dist/full.d.ts",
+      "default": "./dist/full.js"
     }
   },
   "files": [
