+++ src/_internal/formatDate.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert';
-import { createIsHoliday, formatDate } from './core.ts';
+import { formatDate } from './formatDate.ts';
 
 describe('formatDate', () => {
   it('文字列をそのまま返す', () => {
@@ -15,16 +15,5 @@ describe('formatDate', () => {
   it('月・日を 2 桁にゼロパディングする', () => {
     const date = new Date(2025, 0, 5); // 2025-01-05
     assert.strictEqual(formatDate(date), '2025-01-05');
-  });
-});
-
-describe('createIsHoliday', () => {
-  it('lookup.has() を使って判定する', () => {
-    const mockLookup = {
-      has: (key: string) => key === '2025-01-01',
-    };
-    const isHoliday = createIsHoliday(mockLookup);
-    assert.strictEqual(isHoliday('2025-01-01'), true);
-    assert.strictEqual(isHoliday('2025-01-02'), false);
   });
 });
